fix: migration foreign key type and ensure routes

+++ database/migrations/2026_06_03_000000_create_announcements_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->boolean('is_active')->default(false);
             $table->json('target_display');
             $table->timestamp('expires_at')->nullable();
-            $table->unsignedBigInteger('created_by');
+            $table->unsignedInteger('created_by');
             $table->timestamps();
 
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
@@ -28,7 +28,7 @@ return new class extends Migration
 
         Schema::create('announcement_reads', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->unsignedBigInteger('announcement_id');
             $table->timestamp('read_at')->useCurrent();
